requirements.txt
Django==1.7.2
pytz==2019.2
django-crispy-forms
